+++ packages/awino-ui/src/app/menu.tsx
@@ -11,7 +11,7 @@ export const mainMenuLinks: MenuItemType[] = [
     key: 'earn',
     items: [
       { type: 'internal', key: 'deposit', url: '/earn/deposit' },
-      { type: 'internal', key: 'liquidity-staking', url: '/earn/liquidity-staking' },
+      // { type: 'internal', key: 'liquidity-staking', url: '/earn/liquidity-staking' },
       { type: 'internal', key: 'manage-awino', url: '/earn/manage-awino' },
       { type: 'internal', key: 'farms', url: '/earn/farms' },
     ],
@@ -33,8 +33,8 @@ export const mainMenuLinks: MenuItemType[] = [
   },
   {
     type: 'internal',
-    key: 'podl',
-    url: '/podl',
+    key: 'pol',
+    url: '/pol',
   },
   // {
   //   type: 'group',