+++ packages/awino-ui/src/components/pages/earn/liquidity-staking/BriefSection/BriefSection.tsx
@@ -3,11 +3,13 @@ import { Grid, Typography } from '@mui/material';
 import Section from '@/components/layout/Section/Section';
 import StatsItems from '@/components/pages/shared/StatsItems/StatsItems';
 import usePageTranslation from '@/hooks/usePageTranslation';
-import { StatsData } from '@/types/app';
+import { StatsData, StatsFormatter } from '@/types/app';
 
 interface Props {
   items: StatsData;
 }
+
+export const statsFormatters: StatsFormatter[] = [{ value: 'usd' }, { value: 'percent' }, { value: 'usd' }];
 
 export default function BriefSection({ items }: Props) {
   const t = usePageTranslation();
@@ -21,7 +23,7 @@ export default function BriefSection({ items }: Props) {
           <Typography color="text.primary">{t('brief-section.description')}</Typography>
         </Grid>
         <Grid item xs={12} md={8}>
-          <StatsItems items={items} />
+          <StatsItems items={items} formatters={statsFormatters} />
         </Grid>
       </Grid>
     </Section>