+++ packages/awino-ui/src/components/pages/pol/PurchaseSection/TreasurePanel.tsx
@@ -15,7 +15,7 @@ import { formatAmount, formatPercent } from '@/lib/formatters';
 
 import Panel from '../../../general/Panel/Panel';
 
-import { PodlPurchaseData } from './PurchaseSection';
+import { PolPurchaseData } from './PurchaseSection';
 
 const Root = styled(Panel)(({ theme }) => ({
   '.AwiPanel-content': {
@@ -97,7 +97,7 @@ const Root = styled(Panel)(({ theme }) => ({
 }));
 
 interface Props {
-  data: PodlPurchaseData;
+  data: PolPurchaseData;
   treasury: any;
 }