+++ packages/awino-ui/src/pages/market/index.tsx
@@ -8,22 +8,31 @@ import storeWrapper from '@/app/store';
 import Seo from '@/components/layout/Seo/Seo';
 import AssetSection from '@/components/pages/market/AssetSection/AssetSection';
 import StatsSection from '@/components/pages/shared/StatsSection/StatsSection';
+import { StatsFormatter } from '@/types/app';
 
 const stats = [
-  { value: 89.7, subvalue: 24.72 },
-  { value: 89.7, subvalue: 24.72 },
-  { value: 0.27, subvalue: 24.72 },
-  { value: 273.4, subvalue: 52 },
+  { value: 89.7, subValues: [24.72] },
+  { value: 89.7, subValues: [24.72] },
+  { value: 0.27, subValues: [24.72] },
+  { value: 273.4, subValues: [52] },
   { value: 47.4 },
 ];
 
 const assets = { market: 12345678, platform: 1234567.89 };
 
+export const statsFormatters: StatsFormatter[] = [
+  { value: 'amount', subValues: ['usd'] },
+  { value: 'amount', subValues: ['usd'] },
+  { value: 'usd', subValues: ['usd'] },
+  { value: 'amount', subValues: ['percent'] },
+  { value: 'amount' },
+];
+
 const MarketPage: NextPage = () => {
   return (
     <>
       <Seo />
-      <StatsSection items={stats} />
+      <StatsSection items={stats} formatters={statsFormatters} />
       <AssetSection total={assets} />
     </>
   );